Fix bug where the program would crash if fom was enabled at beginning
This bug was because we would register a file to the FOM system at the beginning of
mmap. However, at the end of mmap is a call to "mmap_region" which will unmap
regions that overlap with the new mapping.
The problem occured in the following steps
1) An anonymous mmap call that overlaps with an existing mmap is made
2) A FOM file is registered with the FOM system in the address range
3) mmap_region is called, unmapping the previously existing overlapped range
4) The FOM system would detect the munmap and delete the file associated with
that range, which was the file we just created
5) The program would try to access the new mmap, but the backing file was deleted

This was fixed by only creating the FOM file at the beginning of do_mmap when
it is needed, and registering the file only after mmap_region is called.