mm: Add support for File Based Memory Management
This patch introduces File Based Memory Management (FBMM), which allows for
memory management that are written as filesystems, similar to HugeTLBFS, to
be used transparently by applications.

The steps for using FBMM are the following:
1) Mount the memory management filesystem (MFS)
2) Enable FBMM by writing 1 to /sys/kernel/mm/fbmm/state
3) Set the MFS an application should allocate its memory from by writing
the MFS's mount directory to /proc/<pid>/fbmm_mnt_dir, where <pid> is the
PID of the target process.

To have a process use an MFS for the entirety of the execution, one could
use a shim program that writes /proc/self/fbmm_mount_dir then calls exec
for the target process. We have created such a shim, which can be found
at [1].

Providing this transparency is useful since it allows applications to use
an arbitrary MFS to mange its memory without having to modify that
application.

Writing memory management functionality as an MFS is useful for more easily
prototyping MM functionality and maintaining support for a variety of
different hardware configurations or application needs.

FBMM was originally created as a research project at the University of
Wisconsin-Madison [2].

Signed-off-by: Bijan Tabatabai <btabatabai@wisc.edu>

[1] https://github.com/multifacet/fbmm-workspace/blob/main/bmks/fbmm_wrapper.c
[2] https://www.usenix.org/conference/atc24/presentation/tabatabai