fbmm: Add helper functions for FBMM MM Filesystems
This patch adds four helper functions to simplify the implementation of
MFSs.

fbmm_swapout_folio: Takes a folio to swap out.
fbmm_writepage: An implementation of the address_space_operations.writepage
callback that simplies writes the page to the default swap space.
fbmm_read_swap_entry: Reads the contents of a swap entry into a page.
fbmm_copy_page_range: Copies that page table corresponding to the VMA of
one process into the page table of another. The pages in both processes
are write protected for CoW.

This patch also adds infrastructure for FBMM to support copy on write.
It adds the copy_page_range callback to the super_operations struct, which
is called when the kernel is copying the page tables for a forked process
using FBMM. The fbmm_copy_page_range helper is our base implementation of
this. It also adds logic to FBMM to handle multiple processes sharing. A
forked process will keep the list of FBMM files it depends on for CoW, and
takes a reference to those FBMM files. To ensure one process doesn't free
memory used by other another, FBMM will only free memory from a file if its
reference count is 1.

Signed-off-by: Bijan Tabatabai <btabatabai@wisc.edu>