Add base implementation of an MFS
Mount by running
sudo mount -t BasicMFS BasicMFS -o numpages=<pages> <mntdir>

Where <pages> is the max number of 4KB pages it can use, and <mntdir> is
the directory to mount the filesystem to.

This patch is meant to serve as a reference for the reviewers and is not
intended to be upstreamed.

Signed-off-by: Bijan Tabatabai <btabatabai@wisc.edu>