TieredMMFS: Use correct offset in tieredmmfs_free_range
The tree of allocated pages for a file in TieredMMFS is indexed
by the page offset of the page in the file. However, in tieredmmfs_free_range,
the code was using the absolute offset to search for pages to remove.
This resulted in munmap just not working.

This patch fixes that.